move config stuff to utils
